requirements.txt
streamlit
python-dotenv
openai
groq
pypdf
docx2txt
google-generativeai
